requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiocache==0.11.1
    # via -r requirements.in
aiohttp==3.8.1
    # via
    #   discord-py
    #   travitia-talk
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via aiohttp
certifi==2021.10.8
    # via
    #   httpx
    #   sentry-sdk
chardet==3.0.4
    # via httpx
charset-normalizer==2.0.12
    # via aiohttp
discord.py @ git+https://github.com/Rapptz/discord.py@934ab4151add79a2140c7084679b15b6c5eb6f91
    # via -r requirements.in
edgedb==0.17.3
    # via -r requirements.in
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
googletrans==4.0.0rc1
    # via -r requirements.in
h11==0.9.0
    # via httpcore
h2==3.2.0
    # via httpcore
hpack==3.0.0
    # via h2
hstspreload==2021.12.1
    # via httpx
httpcore==0.9.1
    # via httpx
httpx==0.13.3
    # via googletrans
hyperframe==5.2.0
    # via h2
idna==2.10
    # via
    #   httpx
    #   yarl
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
orjson==3.6.7
    # via -r requirements.in
pillow==9.0.1
    # via -r requirements.in
pink-accents==0.0.4
    # via -r requirements.in
pydantic[dotenv]==1.9.0
    # via -r requirements.in
python-dotenv==0.19.2
    # via pydantic
rfc3986==1.5.0
    # via httpx
sentry-sdk==1.5.7
    # via -r requirements.in
sniffio==1.2.0
    # via
    #   httpcore
    #   httpx
travitia-talk==0.0.1
    # via -r requirements.in
typing-extensions==4.1.1
    # via
    #   edgedb
    #   pydantic
urllib3==1.26.8
    # via sentry-sdk
uvloop==0.16.0
    # via -r requirements.in
yarl==1.7.2
    # via aiohttp
